Move throttle out of utils
