+++ test/utils/utils.test.js
@@ -97,35 +97,4 @@ describe('Utils', () => {
     expect(utils.updateObj({}, allKeys)).to.eql(utils.cloneObj(allKeys));
     expect(utils.updateObj({ a: 'blah', d: 1234 }, allKeys)).to.eql({ a: 'blah', b: 2, c: [6, 7, 8], d: 1234 });
   });
-
-  describe('throttle', () => {
-    let clock;
-
-    before(() => {
-      clock = sinon.useFakeTimers();
-    });
-
-    after(() => {
-      clock.restore();
-    });
-
-    it('callback called twice, once right away and once after delay', () => {
-      const callback = sinon.spy();
-      const throttled = utils.throttle(200, callback);
-      throttled();
-      expect(callback.callCount).to.equal(1);
-      clock.tick(200);
-      expect(callback.callCount).to.equal(2);
-    });
-
-    it('callback only called twice even with multiple throttle calls', () => {
-      const callback = sinon.spy();
-      const throttled = utils.throttle(200, callback);
-      throttled();
-      throttled();
-      throttled();
-      clock.tick(200);
-      expect(callback.calledTwice).to.be.true;
-    });
-  });
 });
